backend/requirements.txt
# Legal AI Analyzer v2.0 — Python Dependencies
#
# Installation:
#   cd backend
#   python -m venv venv
#   venv\Scripts\activate        (Windows)
#   source venv/bin/activate     (Mac/Linux)
#   pip install -r requirements.txt

# ── Core API framework ───────────────────────────────────────────
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# ── PDF processing ───────────────────────────────────────────────
pypdf==3.17.1
fpdf2==2.7.6

# ── Groq API (free LLM — no GPU needed) ─────────────────────────
groq==0.9.0

# ── ML classifier (risk classification + evaluation) ────────────
scikit-learn==1.4.2
numpy==1.26.4
joblib==1.4.2
imbalanced-learn==0.12.3        # class-balanced training utilities

# ── SHAP explainability ──────────────────────────────────────────
# Provides feature-level attribution (which words drove the prediction)
shap==0.45.1

# ── NLP (optional — improves clause extraction and entity detection) ─
# Install then download model:
#   pip install spacy
#   python -m spacy download en_core_web_sm
# spacy==3.7.4
